Изменила структуру
