+++ Lesson_04/src/main/java/kasyanov/task_01/Singleton.java
@@ -1,4 +1,4 @@
-package kasyanov;
+package kasyanov.task_01;
 
 public final class Singleton {
