+++ Lesson_04/src/main/java/kasyanov/task_02/entities/Archer.java
@@ -1,0 +1,5 @@
+package kasyanov.task_02.entities;
+
+public interface Archer {
+    public void shoot();            // Интерфейс  Лучника - Стрелять
+}
