+++ Lesson_04/src/main/java/kasyanov/task_02/entities/ElfArcher.java
@@ -1,0 +1,6 @@
+package kasyanov.task_02.entities;
+
+public class ElfArcher implements Archer {
+    public void shoot() {            // метод: использовать лук (Elf)
+    }
+}
