+++ Lesson_04/src/main/java/kasyanov/task_02/entities/ElfMage.java
@@ -1,0 +1,6 @@
+package kasyanov.task_02.entities;
+
+public class ElfMage implements Mage {
+    public void cast() {             // метод: использовать магию (Elf)
+    }
+}
