+++ Lesson_04/src/main/java/kasyanov/task_02/entities/ElfWarrior.java
@@ -1,0 +1,6 @@
+package kasyanov.task_02.entities;
+
+public class ElfWarrior implements Warrior {
+    public void attack() {           // метод: использовать меч (Elf)
+    }
+}
