+++ Lesson_04/src/main/java/kasyanov/task_02/entities/HumanArcher.java
@@ -1,0 +1,7 @@
+package kasyanov.task_02.entities;
+
+
+public class HumanArcher implements Archer {
+    public void shoot() {            // метод: использовать лук (Human)
+    }
+}
