+++ Lesson_04/src/main/java/kasyanov/task_02/entities/HumanMage.java
@@ -1,0 +1,7 @@
+package kasyanov.task_02.entities;
+
+
+public class HumanMage implements Mage {
+    public void cast() {             // метод: использовать магию (Human)
+    }
+}
