+++ Lesson_04/src/main/java/kasyanov/task_02/entities/HumanWarrior.java
@@ -1,0 +1,6 @@
+package kasyanov.task_02.entities;
+
+public class HumanWarrior implements Warrior {
+    public void attack() {           // метод: использовать меч (Human)
+    }
+}
