+++ Lesson_04/src/main/java/kasyanov/task_02/entities/Mage.java
@@ -1,0 +1,5 @@
+package kasyanov.task_02.entities;
+
+public interface Mage {
+    public void cast();             // Интерфейс  Лучника - Стрелять
+}
