+++ Lesson_04/src/main/java/kasyanov/task_02/entities/Warrior.java
@@ -1,0 +1,6 @@
+package kasyanov.task_02.entities;
+
+
+public interface Warrior {
+    public void attack();           // Интерфейс  Воина - Атаковать
+}
